libdrm: un-break by dropping GPG signature
The used key was missing from drm.gpg, apparently (at least for my gpg
command ... where did Treeve get the key from?!). We discussed this again
and again: Relying on upstream sigs is bad. At least we'd need to also
verify the hash of the sig.

Otherwise, the file can be changed by upstream or attacker with access
to a key at will. And regarding access to a key: We accumulate key after
key and increase the attack surface for compromising one of them.

Thinking again: Maybe the key just expired right now. That is the other
reason: Upstream sigs break at predetermined times whent he keys expire
and gpg refuses to check using them!

DO NOT USE UPSTREAM GPG. We should really purge that. Now.

Upstream signatures should be checked once by the developer doing an
update, locally. Then the hash enters the grimoire. This is stable and
as trustworthy as we can get.

Now we need that spell that installs and updates a database (with wrapper
script) of upstream keys for mages to use.