crypto/openssl: Rework the nasty wget rebuild hack.
Besides the work in PRE_REMOVE, openssl/BUILD also built
and installed a rogue wget binary to work with the to-be updated
openssl libs. It did this in a very crude manner that also
overwrote existing openssl libs before installing.

This even manage to overwrite libssl.so.1.1 with the contents
of libssl.so.3 due to symlinks being followed. A nice way
to kill my SSH daemon and active sessions. Thank you.

Also, the wget build did not work anyway for the upgrade from
1.1 to 3, as it used installed headers to link with to-be-installed lib.

Now wget is built using matching headers and libraries, without
pre-installing any of them, and slipped in in FINAL. It is still
a nasty hack, but more likely to do the right thing now.