build_system_handler.function: normal depends uses
Improve to handle normal (optional_)depends usages where only a single
build system is used for a spell. This enables the same functions to be
used whether a spell supports only a single build system or multiple.
It also enables more succinct code when the current build system matches
the default build system of the spell (i.e., SPELL_BUILD_SYSTEM).

The main difference between the syntax provided by the functions here
and the normal (optional_)depends function calls is the order of the
arguments. However, by making use of default values, the syntactic order
of the normal functions (see [1] below) can be emulated with only
inserting an additional "-o" before the option name (see [2] below).
When a depends call explicitly provides neither the disabled option
value nor the message, the functions here are equal to the normal
functions except for the addition of the "-o" flag and the change from
the full enabled specification to only giving the name of the enabled
option.

1: SPELL OPTION_ENABLED OPTION_DISABLED MESSAGE
2: SPELL -o OPTION MESSAGE ENABLED_OVERRIDE DISABLED_OVERRIDE