graphics-libs/glfw: update to 3.3.9
Update to 3.3.9 and improve option/dependency handling. Provide the user
with clearer choices for the optional functionality to enable. Respect
the user's choice for whether to enable testing.

Currently, enabling examples causes the build to fail, so examples are
forcefully disabled.

Convert depends functions to buildsys_depends to future-proof both if
the spell switches to a different build system(s) in the future, as well
as to make adding new cmake options cleaner. Also, this spell can serve
as a reference for how to use buildsys_depends in lieu of the regular
(optional_)depends functions now that the buildsys_depends functions can
cover those cases as well.