xserver: add option for X11 Security extension
This has been disabled by default upstream, but is still assumed in a default
build of ssh for the -X option as opposed to -Y (ForwardX11Trusted).

You may debate that X11 forwarding is bad and the protections in place by
untrusted forwarding are inadequate ... but better nothing if not perfect?
Let the user decide.