devel/llvm: option of host toolchain for building
Allow the user to choose which host compiler (gcc or llvm) to build
with, depending on which ones are present in the host system at the time
of casting the llvm spell. Also allow choosing among ld (system
default), gold, mold, and lld as the linker to use when building.

Refactored and cleaned up more of the spell files. Corrected dependency
requirements and options. Corrected/updated the license to match the
current license of the sources. Updated the description text.

The DEPENDS file was converted to use the buildsys depends functions to
ease future maintenance and improve readability, and the build system
(cmake) is now declared explicitly in DETAILS.