devel/llvm: fix and improve builds, add more support
Fix the support for SPIR-V and make the target, project, and runtime
selection and usage in BUILD more robust.

Correct CLANG_ to LLVM_ for the setting name that enables/disables Z3
solver support.

Add support for building SPIR-V and Vulkan runners in MLIR.

Explicitly set or unset the option for enabling doxygen documentation in
the build rather than relying on the build system to correctly pick it
up automatically based on the presence or absence of doxygen in the host
system. Explicitly include clang-tools-extra in the projects to build,
and explicitly ensure that the LLVM utilities are installed as well as
built if the user enables them.

Leave the linker choice as-is in PREPARE and manipulate the selection
string in BUILD instead. When the selected linker is lld (or equivalent
to it, even via symlink as the system linker), force the LLVM_ENABLE_LLD
setting to on to ensure that the LLVM build takes advantage of using
lld.