graphics-libs/mesa: add support, improve options
Add support for more drivers. Upstream mesa includes drivers for ARM
platforms, and they are now added here.

Improve option presentation and handling. For options that are entirely
dependent on other options being enabled, they are only presented when
their dependents are enabled. Many of the option messages were also
reworded or expanded upon to better clarify their meaning and use for
the user.

The handling of empty list options was also corrected in BUILD.