graphics-libs/mesa: refactor and fix CONFIGURE
The use of 'auto' for driver options is problematic because the build system
will select drivers, but we have no way to know what it will select, and
thus no way to determine later queries that depend on which drivers are
enabled. Such instances of 'auto' were therefore removed as a choice
even though they are supported in the build system.

Queries that offer 'all' are also problematic, but for a different
reason. In the query, the meaning of 'all' is restricted to the subset
of what is available based on preceding query choices. In the build
system, 'all' means literally all drivers with no restriction to the
subset. Therefore, when a user chooses 'all', we must transform the
values to the currently restricted set rather than directly passing the
value 'all' to the build system.

The handling of persistent variables was corrected to not include
newlines, and to not directly mangle query variables, but to copy their
values and then operate on those persistent copies. Additionally, the
persistent variables are guaranteed to not include newlines, as the
persistent_add function seems to interpret newlines as delineating
separate variables to persist.