gcc: fix support for linker plugins (++PATCHLEVEL)
Detection of linker support for plugins during gcc build was broken since us removing
.la files all around. There's a hardcoded test for a .la plugin file in configure, but a
.so file is there. Somehow we are messing with the gcc build by making it only create

gcc/liblto_plugin.so

instead of also the matching .la.

I patched configure itself to not touch anything else in that nested codebase.  I did
not track down the point where we apparently mess with the build. I guess we break this,
and I am putting a band-aid on.

Result of the broken state was that gcc was built without support for -fuse-linker-plugin,
despite building a plugin. This broke random other builds, like python3 with lto.