simpleinit-msb: fix for 64 bit time_t and safer strncat of utmp line
With GCC 14, the bad usage of time(&ut.ut_time) doesn't fly anymore with
64 bit time_t. This patch fixes that. Still, overflow of the 32 bit data
fields is not handled nicely (errors ignored in simpleinit-msb code).

Also, a strncat() from ut_line did not consider that this could be a
non-null-terminated string. GCC also warned about that.

We still got a use of tmpnam() that we maybe should replace. It is
for naming a fifo. So maybe doing that in a safe temporary directory
would be sensible.