llvm: fix && chain in BUILD, fix build for non-existing llvm-spirv
The call to command -v llvm-spirv causes one if block to return false
if that command is not there. The whole call is there because of some
build system issue that I do not know to apply to us or not. Please
fix that more properly if you know;-)

The missing && caused the build to continue, but the configuration was
off. In my case it meant always building with gcc, not clang.