devel/llvm: improve libclc workaround handling
Explicitly warn the user in PREPARE that selecting libclc SPIR-V targets
will require llvm-spirv to already be present on the system. Also, if
the llvm-spirv command is not available in BUILD, present an error to
the user with a suggestion of how to successfully cast llvm with libclc
SPIR-V support and abort the build (it would later fail anyway).

This workaround is very inelegant, but this bug should only be temporary
upstream, and the present method in the spell avoids the creation of
persistent variables that are only relevant for the temporary workaround
while also providing maximum flexibility to the end user of how to
configure their LLVM.