firefox-esr: version 115.21.0esr, ++SECURITY_PATCH
Note: I was unable to test the build, as my SMGL box is in a state of recovery. I get
a failure from the mach build system early on about its python virtualenv not being
populated. (Btw.: Lots of syntax warnings, too.)

I get the same failure with the previous firefox-esr version that I was able to build
before a number of other system updates and perhaps breakages. Since this update is
fixing a number of security issues, as always, I am not holding it back just because
my machine needs some days or weeks to finally get over an llvm build and other toolchain
recasting.